+++ client/src/components/Background/BackgroundImagesBank.js
@@ -1,0 +1,6 @@
+const backgroundImages = {
+  landingBackground: require("../../assets/pasta-background.jpg"),
+  inputBackground: require("../../assets/skewer-background.png"),
+  resultsBackground: require("../../assets/beef-bowl-background.png"),
+};
+module.exports = backgroundImages;