+++ client/src/components/InputPage/InputPage.jsx
@@ -1,17 +1,10 @@
 import React, { Component } from "react";
-import Background from "../../assets/skewer-background.png";
+import Background from "../Background/Background";
 import Navbar from "../Navbar";
 import Description from "../Description";
 import BrowseImagesButton from "./BrowseImagesButton";
 import SubmitImageButton from "./SubmitImageButton";
 import "./InputPage.css";
-
-const pageBackground = {
-  backgroundImage: `url(${Background})`,
-  backgroundPosition: "center",
-  backgroundSize: "cover",
-  backgroundRepeat: "no-repeat",
-};
 
 class InputPage extends Component {
   constructor(props) {
@@ -54,7 +47,10 @@ class InputPage extends Component {
 
   render() {
     return (
-      <div className="input-page-wrapper" style={pageBackground}>
+      <div className="input-page-wrapper">
+        <div id="input-page-background">
+          <Background backgroundImage="inputBackground" />
+        </div>
         <Navbar />
         <div className="description-wrapper-2">
           <Description descriptionText="First, allow this app to use your location. Then, either take a photo of yourself (and whoever you're with!) or browse your local files for such a photo. Once you upload it, give the app a few seconds, and your restaurant recommendation will be shown!" />